+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 1.  Add constructor for `SaintsDictionary<,>` so it can be created exactly like a standard c-sharp `Dictionary<,>`
 2.  If the key/value for `SaintsDictionary<,>` is a general class/struct which already has a custom drawer, use that custom drawer rather than flat it.
+3.  Fix `SaintsDictionary<,>` can not detect the serialized field when get inherited
 
 ## 3.33.0 ##